# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
      "version": "1.0.0",
      "title": "Swagger Test inapps",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "apiteam@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "none",
    "basePath": "/v2",
...
    "schemes": [
      "https",
      "http"
    ],
    "paths": {
...
      "/api/admin/susspendUser": {
        "post": {
          "tags": [
            "Admin"
          ],
          "summary": "suspenduser",
          "description": "This can only be done by the logged in admin.",
          "operationId": "suspenduser",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json",
            "application/xml"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "id",
              "description": "create admin object",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "suspend user success",
              "schema":
              {
                "type": "object",
              "properties": {
                "qrUrl":{
                "type": "string"}
              }
              }
            },
            "300": {
              "description": "cant find user with id"
            }
          }
        }
      },
...
      "/api/supervisor/upgradeRole": {
        "post": {
          "tags": [
            "Supervisor"
          ],
          "summary": "upgrade role from supervisor",
          "description": "This can only be done by supervisor role.",
          "operationId": "upgrade role",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json",
            "application/xml"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "body",
              "description": "Id from user",
              "required": true,
              "type": "string"
            },
            {
              "name": "role",
              "in": "query",
              "description": "role want to update",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "verify success"
            },
            "300": {
              "description": "cant find user with id"
            }
          }
        }
      },
...
    },
    "securityDefinitions": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      },
      "petstore_auth": {
        "type": "oauth2",
        "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
        "flow": "implicit",
        "scopes": {
          "read:pets": "read your pets",
          "write:pets": "modify pets in your account"
        }
      }
    },
...
  }